feat: User is returned to calling directory after using kafka functions
This commit adds commands in .kafka-functions so that when the user
calls any function in this file, after the function completes, the
user is returned to the directory the called the function from.
Before, the user would go to their home directory, no matter which
directory they were calling from.